requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asn1crypto==0.22.0
Babel==2.5.0
billiard==3.3.0.23
bitstring==3.1.5
bleach==2.0.0
celery==3.1.25
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
decorator==4.1.2
Django==1.8.18
django-celery==3.2.1
django-celery-beat==1.0.1
django-solo==1.1.2
django-suit==0.2.25
entrypoints==0.2.3
flower==0.9.2
geojson==2.1.0
geopy==1.11.0
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
keyring==8.7
keyrings.alt==1.3
kombu==3.0.37
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.3.0
nbformat==4.4.0
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyicloud==0.9.1
pyOpenSSL==17.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
tzlocal==1.4
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.2
